backend/app/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
httpx>=0.27.0
typing-extensions>=4.9.0
